Adding more kitten movement
Now if we don't have 2 kittens on the upper or lower part of the screen they can roam anywhere they want.

This is done by setting either the topLimit or botLimit values in their collision functions.

I also restricted the vertical movement of the top 2 kittens a bit so they don't clip into the box as much.